requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
comtypes==1.4.6
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
newsapi-python==0.2.7
pillow==10.4.0
psutil==6.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyjokes==0.8.3
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.91
pytweening==1.2.0
pywhatkit==5.4
pywin32==306
requests==2.32.3
setuptools==74.1.2
soupsieve==2.6
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wikipedia==1.4.0
